[wip] chore: Renames web3-common to web3-uri/common
